change mpdigit to uint -- 64-bit mp library not so working

+++ include/mp.h
@@ -13,7 +13,7 @@ AUTOLIB(mp)
 
 #define _MPINT 1
 
-typedef ulong mpdigit;
+typedef uint mpdigit;
 
 /* the code assumes mpdigit to be at least an int */
 /* mpdigit must be an atomic type.  mpdigit is defined */
